Holy fucking shit I did it
that was the single most painful thing I've done in rust programming to
date. I needed like 3 pairs of channels just to get the songs to play
continuously wtf. This is probably an AIDS solution but it's what I
could think of given my structure and the whole OutputStream not being
Send while also dying a painful death if it's dropped.

Currenlty all the debug println's are still there and it bricks if you
try to exit but hey it fucking PLAYS so I'M goint to BED.