tui.rs light refactor for basic mouse input
Streamlines certain processes in order to make the dynamic nature of
mouse inputs more manage-able.

TL;DR: Pane enum for fn's that are similar but slightly different
depending if they apply to the queue pane or filter panes. ZoneEvent as
a 'zone'-specific mouse event with the math pre-performed. build_list()
now apart of UI along with new properties and methods to ease layout
construction.