TUI MultiBar™ and search functionality
MultiBar can be drawn in different modes for displaying different
things. It's basically a simplified version of that weird ass context
bar on macbook pros.

The input functions by setting the multibar's draw mode to "input"
which displayes a title and buffer, then simply capturing keys and
pushing chars to the buffer before redrawing (ui.multi_bar_input()).
Finally the resulting input is case-insensitively matched against
the start of the active view's items (ui.search()).

bit weird having to pass the whole library to a simple input getting
method since it has to redraw the whole screen but eh. I toyed with
making UI retain a weak ref to library in the past and didn't like it.