TRACK: TAGSTRING submodule WIP
The TAGSTRING submodule has a basic parser for tagstrings. Can currently
handle substitutions, recursion(not really neeeded but yeah),
and escapes.

Next will be implementing a simple conditional, so you can modify the
output based on whether or not a tag exists.

This is its own submodule within TRACK as it's a parser that needs to
conform to expectations, and therefore contains an entire suite of tests