TUI FilterTreeView widget && new structure (again)
So I gave up on the separate state/display shit cause it was making my
brain cells slowly commit suicide one by one.

FilterTreeView is the first implementation of a "ContainedWidget" that
holds everything it needs to function within itself, minus the
framebuffer which it logically can't. It implements familiar new trait
Clickable with the same process_event(), and new new trait Scrollable,
which provides methods for scrolling the active pane.

Library has some new fns to help with this, and the Filter evt type now
has a bool wrapped stating whether or not the filters were resized. Im
torn on this, as it's also hypothetically possible for FilterTreeView to
figure this out on its own instead of using update_from_library on
resize, but that's to be explored later after cleanup

Theme now just has 8 styles instead of 4 styles and 2 mods.

TUI main has a lot of small rolling changes made as it was broken and
rebuilt with FilterTreeView. Volume keys are now "+-" and 'V' now uses
the fancy new selection fns from FilterTreeView.

Overall, it's 100% going to be less computationally efficient to have
everything query the library for a unique copy of filtered_tracks any
time it needs info on the filters, but the QoL in a coding sense
outweighs the extra few megs of memory used while clicking the ui.
Once Queue (and maybe Statusbar) are migrated to the new system and the
commented code is all removed, it's still possible to re-optimize by
exposing new fns in Library.

tl;dr: this is a lotta BS I worked on over the last few months so it's
very disjointed in thought and structure but I think it's coming
together maybe hopefully come and save me