TUI Queue refactor -> QueueTable ContainedWidget
Also, Queue is now a Table as opposed to a plain List
as a result, LIBRARY now has a new 'sort_tagstring' field which contains
a separate list of tagstrings of which to sort get_queue() by.

"Why?"

There's two ways to have Queue be a sortable table:
 1) leave get_queue() unsorted, and make QueueTable sort it's display
 2) sort the queue before it reaches the table, and simply use the table
    to display it.

The advantages of #1 is that QueueTable can easily sort like you would
any list in your spreadsheet program of choice. This would allow you to
'select' an individual column to sort without manually typign in a new
sort parameter.
The problem with this is the amount of fluff required
for it to look nice: figure out which Track was selected, resort, move
position to Track in new sort. Also any interactions like playing a
track will need to sort, find item, and unsort as well.

The advantages of #2 is there's basically built-in CLI integration, as all
that's necessary is to put the relavant arg into CLAP, sharing the
--filter already in use, in addition to the fact that anything using the
library's .get_queue() method will have the same sort, so no translation
necessary.

I went with #2, as it was by FAR the easiest to implement, easiest to
maintain, and is fitting for OMPL's low-level nature.
In reality (for me), the sorting is mostly a set and done thing, whilst
locating a specific track is best done via the Search method.
Speaking of....

This patch does not yet contain
* a way to modify the sort_tagstrings
* a re-worked search method for the new QueueTable
* a re-worked insert/delete filters.

Ergo, the next major patch in this refactor will have to focus on things
in the Library side, as it shouldn't be up to TUI to implement a filter
insertion on its own.