TUI MenuBar proof-of-concept
Think it works pretty well. Still uses enums instead of FNs directly,
but it's a bit cleaner than ZoneEvent. If there's a way to store FNs in
the Tree, that'd be preferable so the widget can truly be 'contained'
without needing to be manually checked for events.

No code yet for clicking beyond the back button. Don't know if I like
the numbers like that. Will need a way to 'scroll' to see off-screen
items.