BUILD default to Rodio for Windows
https://github.com/RustAudio/cpal/issues/489
https://github.com/RustAudio/cpal/issues/263

Seems you have to manually resample everything yourself on Windows.
Apparently all the modern servers on linux and mac resample it for you
so I'm really not keen on implementing a fallback homemade resampler
which would basically only be used for my VR machine. On the bright side
this means hypothetically Sympal should work on MacOS OOTB. Shame we
dont allow the white devil computers in our house or else I'd test it.