SYMPAL Handle every single error
Every single Result and Option is now either handled or propogated

The only way it can panic now is if cpal or symphonia panic in a weird
way that I don't know I need to catch.

So far it looks like nothing is broken despite the massive diff. No
deadlocks or other strangeness ...yet