Add rudimentary album art decoding
Optional, uses bare `image` crate features

No Display yet. Slow. Either it will have to be lazy decoded or the
meta probe core assignment will need to change into a channel system.

Right now it's only loading the embedded images and it's already causing
a very noticable 300ms startup delay in release. Adding fallback path
searching will be 10x worse.

Currently I just hard resize everything to 32x32 because that's the
maximum reasonable size I will display and the memory usage is small.
I have an idea to hash the images as they come in and store them
in a database to avoid duplicates, but that's a stretch goal.