TRACK mild metadata loading refactor
Shares code between art/tags

I also fixed the weird vorbis/flac mishandling that's been plauging the
app since forever. Turns out when I was trying other solutions and used
the static typed readers as a fix it was actually because the
FormatReader metadata is *different* than the ProbeResult metadata.

So effecitvely the same effect can be achieved file-agnostically by
using the `ProbeResult.format.metadata` as a fallback to
`ProbeResult.metadata`...

Aaaaaaa