LIBRARY Use lock timeouts
Custom trait on RwLock and Mutex to timeout after 2s
of waiting for a guard. Since all the unwraps were defaulted or
otherwise branched in an earlier commit, the behavior of a timeout is so

While this may be a bit confusing if a timeout happens due to no error
propogation, I think it's a huge step up over needing to pkill ompl in
the event of a Library deadlock.